# Conflicts:
#	Rocket.BL.Common/Models/UserRoles/IPermission.cs
#	Rocket.BL.Common/Models/UserRoles/IRole.cs
#	Rocket.BL.Common/Models/UserRoles/Role.cs
#	Rocket.BL.Tests/ReleaseList/FilmDetailedInfoServiceTest.cs
#	Rocket.BL.Tests/ReleaseList/MusicDetailedInfoServiceTest.cs
#	Rocket.BL.Tests/ReleaseList/TVSeriesDetailedInfoServiceTest.cs
#	Rocket.DAL.Common/DbModels/DbUserRole/DbRole.cs
#	Rocket.DAL.Common/Repositories/ReleaseList/IDbPersonRepository.cs
#	Rocket.DAL.Common/UoW/IUnitOfWork.cs

разрезолвил

+++ Rocket.BL.Common/Models/UserRoles/Permission.cs
@@ -7,7 +7,7 @@ namespace Rocket.BL.Common.Models.UserRoles
         /// <summary>
         /// uniq identificator of permission
         /// </summary>
-        public ushort PermissionId { get; set; }
+        public int PermissionId { get; set; }
 
         /// <summary>
         /// description of permission
